Separate sign and kg lib files
