+++ RelayProofsOfConcept/EddsaTendermintServer/mmpc-client/src/lib.rs
@@ -1,3 +1,4 @@
-pub mod eddsa_peer;
+pub mod eddsa_peer_kg;
+pub mod eddsa_peer_sign;
 pub mod peer;
 pub mod tendermint_client;
